Adjusted types in base classes, repos too.

+++ src/main/java/com/vroommates/VroomMates/model/tripmodel/Trip.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class Trip {
     @Id
     @GeneratedValue
-    private Long TripID;
+    private int TripID;
     private int DriverID;
     private boolean IsLive;
